+++ NEWS.md
@@ -6,6 +6,9 @@
 * new functions `get_variable_labels()`, `get_value_labels()`, `get_na_values()`
   and `get_na_range()` identical to `var_label()`, `val_labels()`, `na_values()`
   and `na_range()`, respectively
+* new helpers `label_attribute()`, `get_label_attribute()` and
+  `set_label_attribute()` to manipulate the "label" attribute on any
+  object (#142)
 
 # labelled 2.11.0